Refactor: Extract useSetFlash tests from hooks.spec.jsx
Closes #209
